kill formatting warning
genmodel.c:852: warning: format '%d' expects type 'int', but argument 3 has type 'long int'


git-svn-id: file:///svn/uhexen2/code/trunk@783 eaf649e6-6c13-0410-b194-84221022858e

+++ utils/h2mp_utils/genmodel/genmodel.c
@@ -849,7 +849,7 @@ static void WriteModel (void)
 
 	if (ModelReadIn)
 	{
-		printf ("file size: %d (Saved %d)\n", (int)ftell (modelouthandle), original_size - ftell(modelouthandle) );
+		printf ("file size: %d (Saved %d)\n", (int)ftell (modelouthandle), (int)(original_size - ftell(modelouthandle)) );
 	}
 	else
 	{
